# backend/api_spec.yaml
openapi: "3.0.2"
info:
  title: "WaterProbe API"
  description: "API for waterprobing"
  version: "1"
paths:
...
  /dataStreams:
      get:
        tags: 
        - Datastream
        summary: "Get all datastreams from thing"
        operationId: getDatastreams
        responses:
          200:
            $ref: '#/components/responses/200'
          404:
            $ref: '#/components/responses/404'
                  
        parameters: 
          - $ref: '#/components/parameters/thingRef'
  /dataStream:
    get:
      tags: 
      - Datastream
      summary: "Gets a datastream"
      operationId: getDatastream
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/dataStream'
        
    put:
      tags: 
      - Datastream
      summary: "put to a datastream"
      operationId: putDatastream
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/dataStream'
        
      requestBody:
        description: Optional properties represented as json for the dataStream
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dataStream'
    
    post:
      tags: 
      - Datastream
      summary: "posts a datastream"
      operationId: postDatastream
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
                
      requestBody:
        description: Optional properties represented as json for the dataStream
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dataStream'
    
    delete:
      tags: 
      - Datastream
      summary: "deletes a datastream"
      operationId: deleteDatastream
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/dataStream'
...
  /Location:
    get:
      tags:
       - Location
      summary: "gets a location"
      operationId: getLocation
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
      parameters:
      - $ref: '#/components/parameters/location'
    put:
      tags:
      - Location
      summary: "updates a location"
      operationId: putLocation
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/location'
        
        
    post:
      tags:
       - Location
      summary: "posting a new location"
      operationId: postlocation
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/location'
    
    delete:
      tags: 
      - Location
      summary: "deletes a sensor"
      operationId: deleteLocation
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/location'
...
components:
  responses:
    200:
      description: Successful response
      content:
        application/json:
          schema:
            title: Sample
            type: object
            properties:
              placeholder:
                type: string
                description: Placeholder description
    201:
      description: Successful response
      content:
        application/json:
          schema:
            title: Sample
            type: object
            properties:
              placeholder:
                type: string
                description: Placeholder description
    404:
      description: Not created response
      content:
        text/plain:
          schema:
            title: Thing not cretead
            type: string
            example:
  parameters:
...
    dataStream: 
      name: dataStreamName
      in: query
      description: "Unique dataStreamName"
      schema:
        type: string
...
    location:
      name: locationName
      in: query
      description: "Unique locationName"
      schema:
        type: string
...
    thingRef:
      name: thingRef
      in: query
      description: "Ref to unique thing"
      schema:
        type: string
...
  schemas:
    location:
      type: object
      required: 
        - description
        - encodingType
        - location
      properties:
        description:
          type: string 
        encodingType:
          $ref: '#/components/schemas/valueCode' 
        location:
          type: object 
...
    tmPeriod:
      type: object
...
    valueCode:
      type: object
    gm_envelope:
      type: object
...
    dataStream:
      type: object
      required:
       - name
       - description
       - observationType
       - unitOfMeasurement
       - thingRef
       - sensorRef
       - observedPropertyRef
       
      properties:
        name:
          type: string
        description:
          type: string
        observationType:
          $ref: '#/components/schemas/valueCode'
        unitOfMeasurement:
          type: array	
          items:
            type: object
        observedArea:
          $ref: '#/components/schemas/gm_envelope'
        phenomenonTime:
          $ref: '#/components/schemas/tmPeriod'
        resultTime:
          $ref: '#/components/schemas/tmPeriod'
        thingRef:
          type: string
        sensorRef:
          type: string
        observedPropertyRef:
          type: string
...